Make change I6820f2efbe7458701475cc833d3077022797b2df with qt 5.1

+++ qtlocation/src/plugins/geoservices/nokia/qgeotiledmappingmanagerengine_nokia.cpp
@@ -296,7 +296,7 @@ void QGeoTiledMappingManagerEngineNokia::loadMapVersion()
 
     QJsonObject object = doc.object();
 
-    m_mapVersion.setVersion(object[QLatin1String("version")].toInt());
+    m_mapVersion.setVersion(object[QLatin1String("version")].toDouble());
     m_mapVersion.setVersionData(object[QLatin1String("data")].toObject());
 }
